Disable sending the exit event in the exec-other case
We changed the logic to send exit events immediately instead of sending them after all other events. A result of this change is that sending the exit here will cause the exit to be sent before the child start. This in turn breaks the process tracking in user space.  We intend to account for this in the user-space tracking table, but we are not ready for that yet.